Scanner now returns Ops which evaluate Stack
I wanted to experiement with Object-Oriented programming and tried
something different. I created an Op class which had a single
method 'evaluate'. Evaluate receives a stack and basically
evaluates it.

Before the code looked at each token and based on the type
of the token evaluated some code. In the case of add, sub,
mul, and div tokens this meant popping to values off of the
stack, evaluating a corresponding operation and then placing
it back on the stack.

I moved that logic into the individual Operations. For example,
AddOp overrides the method evaluate from Op. It receives
a stack<Double> as input, pops the two operands off, and
places their sum on the stack. The other operations do
the same thing.

I also created a Num which though a value still peforms an operation.
Num pushes its value on the stack.

I also considered creating another class BinaryOp so that the
repeated code in AddOp, SubOp, MulOp, and DivOp could be reused
but I didn't think it was necessary to add the indirection.

As it is now every Op simply knows how to evaluate the stack.
If I wanted to implement negation, a unary operation, I wouldn't
know ahead of time whether a SubOp is Binary or Unary. I could
just wait until I actually am evaluting the stack to make that
decision. If I created another class for BinaryOp and UnaryOp
I would have to make the decision when scanning the tokens
and that doesn't seem like it would be worth it.